Work through the Pytorch DCGAN tutorial
A basic DGCAN using the CelebA dataset, this is the Pytorch tutorial but
refactored out into classes.

The refactoring is both to aid my own understanding of what's going on
as opposed to blindly copying and pasting from the tutorial, and to
support the future growth of the codebase that the project will require.