Reorder some code, reduce hardcoded numbers in Trainer
Stop hardcoding the feature_map_size and the latent_vector_size in the
Trainer - make it a member variable so we can parameterise it later.

Reorder the code in the constructor a little so all the parameters are
set together.

Improve the logging around initialisation so we see all of the
parameters in the startup log.